Rework flatten_roles
flatten_roles needs input roles:
- for Azure, the roles are only accessible by user_id
- for local-db users, the roles are only accessible by user_name

The previous flatten_roles API only accepted a user_name, which made it
impossible to be used with an Azure user, which is why it had to use the impl_
function. This led to a bug in the introspection methods, because flatten_roles
simply wouldn't work with an Azure user.

Both access_level_for_request_params and the METH_USER_PROFILE/METH_USER_ROLES
methods handle the case, where there are just no roles defined for a particular
peer. I propose a new API:
- user_base_roles fetches whatever is defined for a peer_id. This will be the
  list of roles defined in the config in case of a local-db user, or the list
  mapped Azure groups for an Azure user.
  - in case this is not a local-db/Azure user, the caller can act accordingly
    (this will usually be an error)
- flatten_roles does the role flattening algorithm